Consolidate search helpers
