+++ search.go
@@ -1,6 +1,8 @@
 package goa4web
 
 import (
+	"database/sql"
+	"errors"
 	"log"
 	"net/http"
 	"strings"
@@ -60,6 +62,52 @@ func InsertWordsToLinkerSearch(w http.ResponseWriter, r *http.Request, wordIds [
 			log.Printf("Error: addToLinkerSearch: %s", err)
 			http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
 			return true
+		}
+	}
+	return false
+}
+
+func InsertWordsToImageSearch(w http.ResponseWriter, r *http.Request, wordIds []int64, queries *Queries, pid int64) bool {
+	for _, wid := range wordIds {
+		if err := queries.AddToImagePostSearch(r.Context(), AddToImagePostSearchParams{
+			ImagepostIdimagepost:           int32(pid),
+			SearchwordlistIdsearchwordlist: int32(wid),
+		}); err != nil {
+			log.Printf("Error: addToImagePostSearch: %s", err)
+			http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
+			return true
+		}
+	}
+	return false
+}
+
+func InsertWordsToWritingSearch(w http.ResponseWriter, r *http.Request, wordIds []int64, queries *Queries, wacid int64) bool {
+	for _, wid := range wordIds {
+		if err := queries.AddToForumWritingSearch(r.Context(), AddToForumWritingSearchParams{
+			WritingIdwriting:               int32(wacid),
+			SearchwordlistIdsearchwordlist: int32(wid),
+		}); err != nil {
+			log.Printf("Error: addToForumWritingSearch: %s", err)
+			http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
+			return true
+		}
+	}
+	return false
+}
+
+func InsertWordsToForumSearch(w http.ResponseWriter, r *http.Request, wordIds []int64, queries *Queries, cid int64) bool {
+	for _, wid := range wordIds {
+		if err := queries.AddToForumCommentSearch(r.Context(), AddToForumCommentSearchParams{
+			CommentsIdcomments:             int32(cid),
+			SearchwordlistIdsearchwordlist: int32(wid),
+		}); err != nil {
+			switch {
+			case errors.Is(err, sql.ErrNoRows):
+			default:
+				log.Printf("Error: addToForumCommentSearch: %s", err)
+				http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
+				return true
+			}
 		}
 	}
 	return false
